Removing redundant code
Removed redundant neural network definitions in the Highams2019 code. All code now points to the original Highams 2019 file, or one of the files containing a trained model based on that file. Need to update the PyTorch Hessian and Runtimes in Highams folder. Also need to delete / relocate some of the validation code.